+++ mirrord/agent/src/main.rs
@@ -573,7 +573,7 @@ async fn main() -> Result<()> {
         .with(
             tracing_subscriber::fmt::layer()
                 .with_thread_ids(true)
-                .with_span_events(FmtSpan::ACTIVE)
+                .with_span_events(FmtSpan::NEW | FmtSpan::CLOSE)
                 .compact(),
         )
         .with(tracing_subscriber::EnvFilter::from_default_env())