+++ CHANGELOG.md
@@ -7,6 +7,12 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+### Changed
+
+- Changed `with_span_events` from `FmtSpan::Active` to `FmtSpan::NEW | FmtSpan::CLOSE`.
+  Practically this means we will have less logs on enter/exit to span and only when it's first created
+  and when it's closed.
+
 ## 3.26.1
 
 ### Fixed